# docs/baseline/openapi.yaml
openapi: 3.1.0
info:
  title: Gamja Market API
  version: 0.9.0
  description: Account, product, chat, safety, trade, configurable Toss payment, durable webhook, settlement, and administrator operations contract. Authentication uses an opaque HttpOnly session cookie.
servers:
  - url: /api/v1
paths:
...
  /payments/orders:
    post:
      operationId: createPaymentOrder
      description: Creates a payment order from the accepted trade server-side KRW price.
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 200 } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreatePaymentOrderRequest' } } } }
      responses:
        '201':
          description: Sandbox payment order created, or the first response replayed for the same request and idempotency key.
          content: { application/json: { schema: { $ref: '#/components/schemas/PaymentView' } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /payments/{orderId}/confirm:
    post:
      operationId: confirmPayment
      description: Confirms a provider payment after matching the buyer, path order ID, and server-side amount.
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: orderId, in: path, required: true, schema: { type: string, minLength: 1 } }
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 200 } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ConfirmPaymentRequest' } } } }
      responses:
        '201':
          description: Sandbox payment approved, or the first response replayed for the same request and idempotency key.
          content: { application/json: { schema: { $ref: '#/components/schemas/PaymentView' } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /payments/{orderId}/cancel:
    post:
      operationId: cancelPayment
      description: Cancels an approved provider payment for its buyer while the trade remains accepted.
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: orderId, in: path, required: true, schema: { type: string, minLength: 1 } }
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 200 } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentReasonRequest' } } } }
      responses:
        '201':
          description: Sandbox payment cancelled and settlement paused, or the first response replayed.
          content: { application/json: { schema: { $ref: '#/components/schemas/PaymentView' } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /payments/{orderId}/refund:
    post:
      operationId: refundPayment
      description: Refunds an approved provider payment for a confirmed or disputed participant trade.
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: orderId, in: path, required: true, schema: { type: string, minLength: 1 } }
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 200 } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentReasonRequest' } } } }
      responses:
        '201':
          description: Sandbox payment refunded and settlement paused, or the first response replayed.
          content: { application/json: { schema: { $ref: '#/components/schemas/PaymentView' } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /payments/{orderId}:
    get:
      operationId: getPayment
      description: Returns payment detail only to the trade buyer or seller.
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: orderId, in: path, required: true, schema: { type: string, minLength: 1 } }
      responses:
        '200':
          description: Participant-visible payment detail.
          content: { application/json: { schema: { $ref: '#/components/schemas/PaymentView' } } }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /payments/webhooks/toss:
    post:
      operationId: receiveTossPaymentWebhook
      description: Accepts a signed Toss payment webhook, durably preserving the exact raw JSON body before asynchronous lookup verification and duplicate or sequence handling.
      x-raw-body-required: true
      security: []
      parameters:
        - { name: toss-transmission-id, in: header, required: true, schema: { type: string, minLength: 1 } }
        - { name: toss-transmission-time, in: header, required: true, schema: { type: string, minLength: 1 } }
        - { name: toss-transmission-signature, in: header, required: true, schema: { type: string, minLength: 32 } }
      requestBody:
        required: true
        description: JSON is parsed for processing while its exact raw UTF-8 bytes are retained for the webhook record.
        content: { application/json: { schema: { $ref: '#/components/schemas/TossPaymentWebhookRequest' } } }
      responses:
        '202':
          description: Webhook durably accepted for asynchronous provider verification, or reported as duplicate.
          content: { application/json: { schema: { $ref: '#/components/schemas/PaymentWebhookReceipt' } } }
        '400': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: gm_session }
  schemas:
...
    CreatePaymentOrderRequest:
      type: object
      additionalProperties: false
      required: [tradeId]
      properties:
        tradeId: { type: string, minLength: 1 }
    ConfirmPaymentRequest:
      type: object
      additionalProperties: false
      required: [paymentKey, orderId, amountKrw]
      properties:
        paymentKey: { type: string, minLength: 8, maxLength: 200 }
        orderId: { type: string, minLength: 1 }
        amountKrw: { type: string, pattern: '^[0-9]+$', description: Exact server-priced KRW integer in the minimum unit. }
    PaymentReasonRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: { type: string, minLength: 2, maxLength: 500 }
    TossPaymentWebhookRequest:
      type: object
      additionalProperties: false
      required: [eventType, sequence, data]
      properties:
        eventType: { type: string, minLength: 1 }
        sequence: { type: integer }
        data:
          type: object
          additionalProperties: false
          required: [paymentKey, orderId, amountKrw, status]
          properties:
            paymentKey: { type: string, minLength: 1 }
            orderId: { type: string, minLength: 1 }
            amountKrw: { type: string, pattern: '^[0-9]+$' }
            status: { type: string, enum: [DONE, CANCELED, REFUNDED] }
    PaymentView:
      type: object
      additionalProperties: false
      required: [id, tradeId, orderId, paymentKey, amountKrw, status, settlementStatus, methodMasked, approvedAt, cancelledAt, refundedAt]
      properties:
        id: { type: string }
        tradeId: { type: string }
        orderId: { type: string }
        paymentKey: { type: [string, 'null'] }
        amountKrw: { type: string, pattern: '^[0-9]+$' }
        status: { type: string, enum: [READY, APPROVED, UNCONFIRMED, FAILED, CANCEL_PENDING, CANCELLED, REFUND_PENDING, REFUNDED] }
        settlementStatus: { type: string, enum: [HOLD, PAUSED, READY, PROCESSING, SETTLED, FAILED, SANDBOX_SETTLED] }
        methodMasked: { type: [string, 'null'] }
        approvedAt: { type: [string, 'null'], format: date-time }
        cancelledAt: { type: [string, 'null'], format: date-time }
        refundedAt: { type: [string, 'null'], format: date-time }
        settlementAvailableAt: { type: [string, 'null'], format: date-time }
        settledAt: { type: [string, 'null'], format: date-time }
        failureCode: { type: [string, 'null'] }
    PaymentWebhookReceipt:
      type: object
      additionalProperties: false
      required: [accepted, duplicate, status]
      properties:
        accepted: { type: boolean, const: true }
        duplicate: { type: boolean }
        status: { type: string, enum: [RECEIVED, PROCESSED, IGNORED, FAILED] }
    Problem:
      type: object
      required: [type, title, status, code, requestId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        requestId: { type: string }
  responses:
    Problem:
      description: Safe problem response.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }